+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change Log
 v1.3.1 (2022.08.xx)
 -------------------
 - Add variants to media
+- Add alt_text to media
 - Bump Kotlin 1.7.10
 
 v1.3.0 (2022.07.05)